Logs not downloading fixed
The previous code relied on the existence of a ‘BatchId’ to determine
which log to download. This does not appear to exist anymore.

The change is to download the latest log file. Also, the name
of the log file is simply based on the date at the time of retrieval.